Moved view for adding reviews to the file which has all views for reviews
