+++ beauty/api/views/review_views.py
@@ -2,14 +2,15 @@
 
 import logging
 
-from rest_framework import (status, filters)
+from rest_framework import (filters, status)
 from rest_framework.generics import (GenericAPIView, RetrieveAPIView)
 
 from rest_framework.response import Response
+from rest_framework.permissions import IsAuthenticated
 
-from api.models import Review
+from api.models import (CustomUser, Review)
 
-from api.serializers.review_serializers import ReviewDisplaySerializer
+from api.serializers.review_serializers import (ReviewAddSerializer, ReviewDisplaySerializer)
 
 
 logger = logging.getLogger(__name__)
@@ -43,3 +44,37 @@ class ReviewDisplayDetailView(RetrieveAPIView):
     """View for retrieving specific review."""
     queryset = Review.objects
     serializer_class = ReviewDisplaySerializer
+
+
+class ReviewAddView(GenericAPIView):
+    """Create Review view.
+
+    This class represents a view which is accessed when someone
+    is trying to create a new Review. It makes use of the POST method,
+    other methods are not allowed in this view.
+    """
+
+    serializer_class = ReviewAddSerializer
+    permission_classes = [IsAuthenticated]
+
+    def post(self, request, user):
+        """This is a POST method of the view."""
+        serializer = ReviewAddSerializer(data=request.data)
+        author = self.request.user
+        to_user = CustomUser.objects.get(pk=user)
+        if serializer.is_valid():
+            serializer.save(
+                from_user=author,
+                to_user=to_user,
+            )
+            logger.info(
+                f"User {author} (id = {author.id}) posted a review for"
+                f"{to_user} (id = {to_user.id})",
+            )
+            return Response(status=status.HTTP_201_CREATED)
+        else:
+            logger.info(
+                "Error validating review: "
+                f"Field {serializer.errors.popitem()}",
+            )
+            return Response(status=status.HTTP_400_BAD_REQUEST)
